* Flush st7789v via dma

+++ components/platform/hosal/bl808_e907_hal/hosal_spi.c
@@ -219,7 +219,7 @@ int hosal_spi_finalize(hosal_spi_dev_t *spi)
     return 0;
 }
 
-int hosal_spi_send(hosal_spi_dev_t *spi, const uint8_t *data, uint16_t size, uint32_t timeout)
+int hosal_spi_send(hosal_spi_dev_t *spi, const uint8_t *data, uint32_t size, uint32_t timeout)
 {
     int ret;
